moving user to app router and fixed update user issue
