+++ apps/client/components/userProfile/edit-profile-modal.tsx
@@ -29,7 +29,7 @@ export default function EditProfileModal({
   IPriorities,
   IAvatar,
   fetchUser,
-  closeModal
+  closeModal,
 }: EditProfileModalProps) {
   const openRef = useRef<() => void>(null);
   const { updateUser } = useUser();
@@ -41,8 +41,11 @@ export default function EditProfileModal({
   const [name, setName] = useState<string>(IName);
   const [priorities, setPriorities] = useState<string[]>(IPriorities);
 
-  const { mutateAsync: updateUserProfile, isLoading: updatingProfile } =
-    useMutation(updateUser);
+  const {
+    mutateAsync: updateUserProfile,
+    isLoading: updatingProfile,
+    error: updatingError,
+  } = useMutation(updateUser);
   const [uploadingImage, setUploadingImage] = useState(false);
 
   const { uploadImages } = usePortalProposal();
@@ -54,41 +57,38 @@ export default function EditProfileModal({
   };
 
   const handleUploadImages = async () => {
+    if (profileImage.length === 0) return `${IAvatar}`;
     const newImages = await uploadImages(profileImage);
     return newImages;
   };
 
   const handleUpdateProfile = async () => {
-    try {
-      const userName = appUser?.username;
-      if (!userName) {
-        toast.error('User name is not defined');
-        return;
-      }
-      setUploadingImage(true);
-      const profileImageAvatar = await handleUploadImages();
-      setUploadingImage(false);
-      console.log(profileImageAvatar, 'profileImageAvatar');
-      await updateUserProfile({
-        userName,
-        name,
-        priorities,
-        proficiencies,
-        bio,
-        avatar: profileImageAvatar,
-      });
-      fetchUser();
-      closeModal();
-    } catch (e) {
-      console.log(e);
+    const userName = appUser?.username;
+    if (!userName) {
+      toast.error('User name is not defined');
+      return;
     }
+    setUploadingImage(true);
+    const profileImageAvatar = await handleUploadImages();
+    setUploadingImage(false);
+    // console.log(profileImageAvatar, 'profileImageAvatar');
+    await updateUserProfile({
+      userName,
+      name,
+      priorities,
+      proficiencies,
+      bio,
+      avatar: profileImageAvatar,
+    });
+    fetchUser();
+    closeModal();
   };
 
   const submit = () => {
     toast.promise(handleUpdateProfile(), {
       loading: 'Updating profile',
       success: 'Profile updated',
-      error: 'Error updating profile',
+      error: `Error updating profile`,
     });
   };
 
@@ -115,7 +115,6 @@ export default function EditProfileModal({
       </Dropzone>
       <TextInput
         label="Name"
-        data-autoFocus
         placeholder="Whizzy"
         mt="md"
         value={name}
@@ -176,8 +175,12 @@ export default function EditProfileModal({
         <Button
           my="md"
           className="w-full"
-          variant="outline"
+          variant="subtle"
           disabled={updatingProfile || uploadingImage}
+          onClick={() => {
+            console.log('clicked');
+            closeModal();
+          }}
         >
           Cancel
         </Button>
